requirements.txt
pandas>=1.5.1
biopython>=1.79
